+++ neo/framework/Zip.cpp
@@ -157,7 +157,7 @@ bool idZipBuilder::Update( const char* zipPath, const char* folder, bool cleanFo
 idZipBuilder::GetFileTime
 ========================
 */
-bool idZipBuilder::GetFileTime( const idStr& filename, unsigned long* dostime ) const
+bool idZipBuilder::GetFileTime( const idStr& filename, unsigned int* dostime ) const
 {
 	// RB: FIXME
 #if defined(_WIN32)
