+++ neo/sys/win32/win_cpu.cpp
@@ -130,7 +130,7 @@ double Sys_ClockTicksPerSecond() {
 			}
 			RegCloseKey( hKey );
 			if ( ret == ERROR_SUCCESS ) {
-				ticks = (double) ((unsigned long)ProcSpeed) * 1000000;
+				ticks = (double) ((uintptr_t)ProcSpeed) * 1000000;
 			}
 		}
 	}
@@ -393,8 +393,8 @@ LogicalProcPerPhysicalProc
 // RB: no checks on Win64
 #if !defined(_WIN64)
 #define NUM_LOGICAL_BITS   0x00FF0000     // EBX[23:16] Bit 16-23 in ebx contains the number of logical
-                                          // processors per physical processor when execute cpuid with 
-                                          // eax set to 1
+										  // processors per physical processor when execute cpuid with 
+										  // eax set to 1
 static unsigned char LogicalProcPerPhysicalProc() {
 	unsigned int regebx = 0;
 	__asm {
@@ -414,8 +414,8 @@ GetAPIC_ID
 // RB: no checks on Win64
 #if !defined(_WIN64)
 #define INITIAL_APIC_ID_BITS  0xFF000000  // EBX[31:24] Bits 24-31 (8 bits) return the 8-bit unique 
-                                          // initial APIC ID for the processor this code is running on.
-                                          // Default value = 0xff if HT is not supported
+										  // initial APIC ID for the processor this code is running on.
+										  // Default value = 0xff if HT is not supported
 static unsigned char GetAPIC_ID() {
 	unsigned int regebx = 0;
 	__asm {
@@ -432,7 +432,7 @@ static unsigned char GetAPIC_ID() {
 CPUCount
 
 	logicalNum is the number of logical CPU per physical CPU
-    physicalNum is the total number of physical processor
+	physicalNum is the total number of physical processor
 	returns one of the HT_* flags
 ================
 */
@@ -476,7 +476,7 @@ int CPUCount( int &logicalNum, int &physicalNum ) {
 
 		while( i < logicalNum ) {
 			i *= 2;
- 			PHY_ID_MASK  <<= 1;
+			PHY_ID_MASK  <<= 1;
 			PHY_ID_SHIFT++;
 		}
 		
@@ -511,10 +511,10 @@ int CPUCount( int &logicalNum, int &physicalNum ) {
 			}
 			dwAffinityMask = dwAffinityMask << 1;
 		}
-	        
+			
 		// Reset the processor affinity
 		SetProcessAffinityMask( hCurrentProcessHandle, dwProcessAffinity );
-	    
+		
 		if ( logicalNum == 1 ) {  // Normal P4 : HT is disabled in hardware
 			statusFlag = HT_DISABLED;
 		} else {
@@ -621,9 +621,9 @@ DWORD CountSetBits( ULONG_PTR bitMask ) {
 typedef BOOL (WINAPI *LPFN_GLPI)( PSYSTEM_LOGICAL_PROCESSOR_INFORMATION, PDWORD );
 
 enum LOGICAL_PROCESSOR_RELATIONSHIP_LOCAL {
-    localRelationProcessorCore,
-    localRelationNumaNode,
-    localRelationCache,
+	localRelationProcessorCore,
+	localRelationNumaNode,
+	localRelationCache,
 	localRelationProcessorPackage
 };
